Factor callback out [full ci]
