+++ features/fixtures/mazerunner/jvm-scenarios/src/main/java/com/bugsnag/android/mazerunner/AnrHelper.kt
@@ -2,6 +2,7 @@ package com.bugsnag.android.mazerunner
 
 import android.os.Handler
 import android.os.Looper
+import com.bugsnag.android.OnErrorCallback
 
 val mutex = Any()
 
@@ -30,4 +31,20 @@ fun createDeadlock() {
         },
         1000
     )
+}
+
+// Do not allow system generated ANRs to be sent to Maze Runner
+val filterSystemAnrs = OnErrorCallback { event ->
+    val error = event.errors.first()
+    val method1 = "android.os.BinderProxy.transact"
+    val method2 = "android.app.IActivityManager\$Stub\$Proxy.handleApplicationCrash"
+    if (error.errorClass.equals("ANR") &&
+        error.stacktrace.any { frame -> frame.method.equals(method1) } &&
+        error.stacktrace.any { frame -> frame.method.equals(method2) }
+    ) {
+        CiLog.info("Filtering system generated ANR")
+        false
+    } else {
+        true
+    }
 }
