+++ features/fixtures/mazerunner/jvm-scenarios/src/main/java/com/bugsnag/android/mazerunner/scenarios/LoadConfigurationKotlinScenario.kt
@@ -6,7 +6,7 @@ import com.bugsnag.android.Configuration
 import com.bugsnag.android.EndpointConfiguration
 import com.bugsnag.android.OnErrorCallback
 import com.bugsnag.android.ThreadSendPolicy
-import com.bugsnag.android.mazerunner.CiLog
+import com.bugsnag.android.mazerunner.filterSystemAnrs
 import java.lang.RuntimeException
 import java.util.regex.Pattern
 
@@ -46,21 +46,7 @@ internal class LoadConfigurationKotlinScenario(
                 true
             }
         )
-        // Do not allow system generated ANRs to be sent to Maze Runner
-        testConfig.addOnError { event ->
-            val error = event.errors.first()
-            val method1 = "android.os.BinderProxy.transact"
-            val method2 = "android.app.IActivityManager\$Stub\$Proxy.handleApplicationCrash"
-            if (error.errorClass.equals("ANR") &&
-                error.stacktrace.any { frame -> frame.method.equals(method1) } &&
-                error.stacktrace.any { frame -> frame.method.equals(method2) }
-                ) {
-                CiLog.info("Filtering system generated ANR")
-                false
-            } else {
-                true
-            }
-        }
+        testConfig.addOnError(filterSystemAnrs)
 
         measureBugsnagStartupDuration(this.context, testConfig)
     }
